move command executor out
